+++ client/src/features/theme/index.js
@@ -1,3 +1,0 @@
-export { ThemeProvider } from './components/ThemeProvider';
-export { default as useTheme } from './hooks/useTheme';
-export * from './utils/tokens';